LPD-93208 Move main entry point to its own folder
