+++ modules/apps/frontend-js/frontend-js-audiences-web/node-scripts.config.js
@@ -4,5 +4,5 @@
  */
 
 module.exports = {
-	main: './src/main/resources/META-INF/resources/index.ts',
+	main: './src/main/resources/META-INF/resources/main/index.ts',
 };
