+++ modules/apps/frontend-js/frontend-js-audiences-web/src/main/resources/META-INF/resources/main/detection/util.ts
