+++ modules/apps/frontend-js/frontend-js-audiences-web/test/detection.test.ts
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-import * as audiences from '../src/main/resources/META-INF/resources/implementation';
+import * as audiences from '../src/main/resources/META-INF/resources/main/implementation';
 
 import type {
 	Attribute,
@@ -11,7 +11,7 @@ import type {
 	Conjunction,
 	Operator,
 	Rule,
-} from '../src/main/resources/META-INF/resources/index';
+} from '../src/main/resources/META-INF/resources/main/index';
 
 const URL = 'https://example.com/audiences.json';
